moved loginOrLogout.tsx to the root of the header folder
