+++ frontend/src/layout/header/loginOrLogout.tsx
@@ -1,7 +1,7 @@
 import { Link } from "react-router-dom";
 import cn from "classnames";
-import deleteCookie from "../../../../helpers/deleteCookie";
-import useCurrentUser from "../../../../hooks/useCurrentUser";
+import deleteCookie from "../../helpers/deleteCookie";
+import useCurrentUser from "../../hooks/useCurrentUser";
 
 type Props = {
   handleOpenStatus: () => void;
@@ -32,8 +32,8 @@ const LoginOrLogout: React.FC<Props> = ({ handleOpenStatus }) => {
         to={isAuth ? "/" : "/login"}
         onClick={handleClick}
         className={cn(
-          "rounded-md text-black font-bold px-3 text-lg h-fit py-0.5 hover:text-black",
-          isAuth ? "bg-red-400" : "bg-blue-400"
+          "rounded-md text-black font-bold px-3 text-lg h-fit py-0.5 transition-opacity hover:text-black",
+          isAuth ? "bg-red-400 hover:bg-red-400/70" : "bg-blue-400 hover:bg-blue-400/70"
         )}
       >
         {isAuth ? "Log out" : "Log in"}
